requirements.txt
Django>=3.2
djangorestframework
djangorestframework-simplejwt
django-filter
psycopg2  # 如果使用 PostgreSQL
reportlab